requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
coverage==5.1
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.14.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.9
more-itertools==8.3.0
oauthlib==3.1.0
packaging==20.3
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==5.4.2
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.1.9
